Package structure: moved support files.
Part of #963.
