+++ src/main/java/de/dennisguse/opentracks/chart/TrackDataHubInterface.java
@@ -1,10 +1,10 @@
-package de.dennisguse.opentracks;
+package de.dennisguse.opentracks.chart;
 
 import de.dennisguse.opentracks.content.TrackDataHub;
 
 /**
  * Interface for communication between activities that use {@link de.dennisguse.opentracks.content.TrackDataHub} and their fragments that need thi data hub.
  */
-public interface TrackActivityDataHubInterface {
+public interface TrackDataHubInterface {
     TrackDataHub getTrackDataHub();
 }
